requirements.txt
ultralytics>=8.0.0
cvzone>=1.5.6
opencv-python>=4.5.0
numpy>=1.20.0
mysql-connector-python
python-dotenv>=0.19.0
langchain_google_genai>=0.0.3
# Web application dependencies
flask>=2.0.0
flask-socketio>=5.3.0
pillow>=9.0.0
requests>=2.25.0